Allow parsing of reports where URI scheme is 'file'.
This permits a user to file CPANtesters reports where 'cpanm' has been
targeted to a file on disk -- typically, a file in a minicpan repository.

Add a build.log in which the user targeted a minicpan tarball and in which
'cpanm' identified an uninstalled prerequisite for which it went out to CPAN.
Thus, this build.log demonstrates use of both the 'file' and 'http" schemes.
Use this log as the basis for a new test file, with mocking sub modified
appropriately.

Add 2 entries to MANIFEST.SKIP and re-make manifest.